+++ .github/release-notes.md
@@ -1,1 +1,2 @@
 * Fixed an error when searing for players named "Infinity"
+* Fixed casual selection manager not working